move CalibrateAZFP to a separate script
